workers/whisper/requirements.txt
openai-whisper
boto3
psycopg2-binary
redis
